requirements.txt
tensorflow-gpu<=1.15
keras<2.3
sklearn
tqdm
xopen
keras_bert